Fixed FB email capture. Added FB Profile pic element.
Currently, I’m just doing this using inline styles because we had built
the CSS structure as not using an img DOM element for the profile
picture.  So the easiest way right now is to implement inline CSS
background-image property for the Jade variable.  If we have time
later, we should implement a DOM element method.